Removed test files.
