+++ angular.json
@@ -7,7 +7,20 @@
       "projectType": "application",
       "schematics": {
         "@schematics/angular:component": {
-          "style": "scss"
+          "style": "scss",
+          "skipTests": true
+        },
+        "@schematics/angular:class": {
+          "skipTests": true
+        },
+        "@schematics/angular:directive": {
+          "skipTests": true
+        },
+        "@schematics/angular:pipe": {
+          "skipTests": true
+        },
+        "@schematics/angular:service": {
+          "skipTests": true
         }
       },
       "root": "",
@@ -87,10 +100,10 @@
               "browserTarget": "mapmyproperty:build:production"
             }
           },
-           "assets": [
-              "src/favicon.ico",
-              "src/assets"
-            ]
+          "assets": [
+            "src/favicon.ico",
+            "src/assets"
+          ]
         },
         "extract-i18n": {
           "builder": "@angular-devkit/build-angular:extract-i18n",
@@ -141,6 +154,7 @@
           }
         }
       }
-    }},
+    }
+  },
   "defaultProject": "mapmyproperty"
 }
