+++ arcgis.webpack.config.js
@@ -2,7 +2,11 @@ const ArcGISPlugin = require("@arcgis/webpack-plugin");
 
 module.exports = {
   plugins: [
-    new ArcGISPlugin({}),
+    new ArcGISPlugin({
+      features: {
+        '3d': false
+      }
+    }),
 
   ],
   node: {
