+++ src/app/drawmyranch/esrimap/esrimap.component.html
@@ -1,6 +1,6 @@
 <div #sidebar class="sidebarBlanket">
   <div class="sidebar">
-    <app-sidebar (startDrawing)="startDrawing($event)" [selectedGraphics]="selectedGraphics"></app-sidebar>
+    <!-- <app-sidebar (startDrawing)="startDrawing($event)" [selectedGraphics]="selectedGraphics"></app-sidebar> -->
   </div>
 
 </div>
