+++ src/app/drawmyranch/sidebar/sidebar.component.html
@@ -1,44 +1,1 @@
-<mat-expansion-panel [expanded]="true" color="black">
-  <mat-expansion-panel-header>
-    <mat-panel-title>
-      Draw
-    </mat-panel-title>
-
-  </mat-expansion-panel-header>
-
-  <div>
-    <button mat-raised-button class="graphicsControls1" (click)="startDrawingGraphics('polygon')">Polygon</button>
-    <br>
-    <div style="display: flex;">
-      <div style="flex-grow: 1;">
-        Line Style:<br>
-
-        <mat-select [(ngModel)]="lineStyle" class="matSelect" (selectionChange)="changeGraphicsStyle()">
-          <mat-select-trigger>
-            <svg-icon *ngIf="lineStyle !== 'none'" [src]="'assets/icons/symbols-sls-'+lineStyle+'.svg'"
-              [svgStyle]="{ 'width.px':150, 'fill':lineColor }"></svg-icon>
-            <span *ngIf="lineStyle === 'none'">None</span>
-          </mat-select-trigger>
-          <mat-option *ngFor="let style of lineStyles | keyvalue" [value]="style.value">
-            <svg-icon *ngIf="style.value !== 'none'" [src]="'assets/icons/symbols-sls-'+style.value+'.svg'"
-              [svgStyle]="{ 'width.px':150, 'fill':lineColor }"></svg-icon>
-            <span *ngIf="style.value === 'none'">None</span>
-          </mat-option>
-        </mat-select>
-
-      </div>
-      <div>
-        <app-color-picker [color]="lineColor" [opacity]="lineOpacity" (colorSelected)="changeColor($event)">
-        </app-color-picker>
-      </div>
-    </div>
-    <div>
-      Width<br>
-      0 &nbsp;<mat-slider class="widthSlider mat-slider-has-ticks" [thumbLabel]="true" step="1" tickInterval="1"
-        (change)="changeGraphicsStyle()" [(ngModel)]="lineWidth" min="0" max="10">
-      </mat-slider> &nbsp;10
-    </div>
-
-  </div>
-  <!-- <div style="background: red; color:white">Works almost. not perfect !!!. Wait a day or so</div> -->
-</mat-expansion-panel>
+Sidebar for DMR
