+++ src/app/mapmyproperty/esrimap/esrimap.component.scss
@@ -1,6 +1,6 @@
-@import url("https://js.arcgis.com/4.14/esri/css/main.css");
-@import "src/app/shared/styles/BodyFrame.scss";
-@import "src/app/shared/styles/themes/mmp.theme.scss";
+@import url('https://js.arcgis.com/4.14/esri/css/main.css');
+@import 'src/app/shared/styles/BodyFrame.scss';
+@import 'src/app/shared/styles/palettes/mmp.theme.palettee.scss';
 
 $translucent: mat-color($mmp-primary, translucent);
 $accent: mat-color($mmp-accent);
