+++ src/app/mapmyproperty/sidebar/sidebar.module.ts
@@ -1,0 +1,10 @@
+import { SidebarComponent } from './sidebar.component';
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+@NgModule({
+  declarations: [SidebarComponent],
+  imports: [CommonModule],
+  exports: [SidebarComponent],
+})
+export class SidebarModule {}
