+++ src/app/shared/styles/themes.scss
@@ -1,23 +1,23 @@
-@import "~@angular/material/theming";
-@include mat-core();
-@import "./themes/mmp.theme";
+// @import "~@angular/material/theming";
+// @include mat-core();
+// @import "./themes/mmp.theme";
 
-@import "./functions";
-@import "./_variables";
-@import "./layout";
+// @import "./functions";
+// @import "./_variables";
+// @import "./layout";
 
-$mmp-light-theme: mat-light-theme($mmp-primary, $mmp-accent, $mmp-warn);
-$mmp-foreground: my-mat-theme-foreground(mat-color($mmp-primary, 700));
+// $mmp-light-theme: mat-light-theme($mmp-primary, $mmp-accent, $mmp-warn);
+// $mmp-foreground: my-mat-theme-foreground(mat-color($mmp-primary, 700));
 
-$mmp-theme: map-merge(
-  $mmp-light-theme,
-  (
-    foreground: $mmp-foreground,
-  )
-);
+// $mmp-theme: map-merge(
+//   $mmp-light-theme,
+//   (
+//     foreground: $mmp-foreground,
+//   )
+// );
 
-// $pmlo-theme: mat-light-theme($pmlo-primary, $pmlo-accent, $pmlo-warn);
+// // $pmlo-theme: mat-light-theme($pmlo-primary, $pmlo-accent, $pmlo-warn);
 
-.mmp-theme {
-  @include angular-material-theme($mmp-theme);
-}
+// .mmp-theme {
+//   @include angular-material-theme($mmp-theme);
+// }
