+++ src/app/shared/styles/variables/mmp.theme.colors.scss
@@ -1,0 +1,66 @@
+$mmpPrimaryPalettee: (
+  50: #eceee3,
+  100: #d0d6b8,
+  200: #b0ba89,
+  300: #909e5a,
+  400: #798a36,
+  500: #617513,
+  600: #596d11,
+  700: #4f620e,
+  800: #45580b,
+  900: #334506,
+  translucent: rgba(97, 117, 19, 0.5),
+  A100: #d4ff7b,
+  A200: #c3ff48,
+  A400: #b2ff15,
+  A700: #a8fa00,
+  contrast: (
+    50: #000000,
+    100: #000000,
+    200: #000000,
+    300: #000000,
+    400: #ffffff,
+    500: #ffffff,
+    600: #ffffff,
+    700: #ffffff,
+    800: #ffffff,
+    900: #ffffff,
+    A100: #000000,
+    A200: #000000,
+    A400: #000000,
+    A700: #000000,
+  ),
+);
+
+$mmpSecondaryPalettee: (
+  50: #e7e7e7,
+  100: #c2c2c2,
+  200: #9a9a9a,
+  300: #727272,
+  400: #535353,
+  500: #353535,
+  600: #303030,
+  700: #282828,
+  800: #222222,
+  900: #161616,
+  A100: #ef6f6f,
+  A200: #ea4141,
+  A400: #f80000,
+  A700: #df0000,
+  contrast: (
+    50: #000000,
+    100: #000000,
+    200: #000000,
+    300: #ffffff,
+    400: #ffffff,
+    500: #ffffff,
+    600: #ffffff,
+    700: #ffffff,
+    800: #ffffff,
+    900: #ffffff,
+    A100: #000000,
+    A200: #ffffff,
+    A400: #ffffff,
+    A700: #ffffff,
+  ),
+);
